kafka: extract topic configs code
This extracts the topic config specific code in describe_configs and
moves it into a separate file so that it can later be reused in the
CreateTopic handler.

This commit only moves the code, the only changes are to make the commit
compile:
 * making the methods in the new file "static inline"
 * creating the report_topic_config method to wrap the extracted code
   that used to be inline in describe_configs.
